backend/requirements.txt
annotated-types==0.7.0
anyascii==0.3.2
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.3
blis==1.2.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
contractions==0.1.73
cymem==2.0.11
dill==0.3.9
Django==5.1.7
django-cors-headers==4.7.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
docling==2.28.0
docling-core==2.23.3
docling-ibm-models==3.4.1
docling-parse==4.0.0
easyocr==1.7.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
filelock==3.18.0
filetype==1.2.0
fsspec==2025.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
huggingface-hub==0.29.3
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.4.2
jsonlines==3.1.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langcodes==3.5.0
language_data==1.3.0
latex2mathml==3.77.0
lazy_loader==0.4
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
marko==2.1.2
MarkupSafe==3.0.2
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.17
murmurhash==1.0.12
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
numpy==2.2.4
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pickle-mixin==1.0.2
pillow==11.1.0
pluggy==1.5.0
preshed==3.0.9
proto-plus==1.26.1
protobuf==5.29.4
pyahocorasick==2.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
pylatexenc==2.10
PyMuPDF==1.25.4
pyparsing==3.2.3
pypdfium2==4.30.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2025.1
pywin32==310
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
rtree==1.4.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
semchunk==2.2.2
sentence-transformers==3.4.1
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
sympy==1.13.1
tabulate==0.9.0
textsearch==0.0.24
thinc==8.3.4
threadpoolctl==3.6.0
tifffile==2025.3.13
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.50.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
word2number==1.1
wrapt==1.17.2
XlsxWriter==3.2.2
